# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A high-performance Git version control system API with memory-first architecture",
        "title": "GOVC API",
        "termsOfService": "https://github.com/caiatech/govc/blob/main/LICENSE",
        "contact": {
            "name": "GOVC API Support",
            "url": "https://github.com/caiatech/govc/issues",
            "email": "support@caiatech.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/caiatech/govc/blob/main/LICENSE"
        },
        "version": "1.0.0-dev"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/repositories/{repo_id}/branches": {
            "get": {
                "description": "Retrieves a list of all branches in the repository",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Git Operations"
                ],
                "summary": "List all branches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of branches",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "branches": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/api.BranchResponse"
                                    }
                                },
                                "count": {
                                    "type": "integer"
                                },
                                "current": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new branch from the current or specified branch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Git Operations"
                ],
                "summary": "Create a new branch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Branch creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateBranchRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Branch created successfully",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repository or source branch not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Branch already exists",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/repositories/{repo_id}/branches/{branch}": {
            "delete": {
                "description": "Permanently deletes a branch (cannot delete current branch)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Git Operations"
                ],
                "summary": "Delete a branch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Branch name",
                        "name": "branch",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Cannot delete current branch",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/repositories/{repo_id}/checkout": {
            "post": {
                "description": "Switches the working directory to a different branch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Git Operations"
                ],
                "summary": "Checkout a branch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Checkout request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CheckoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch checked out successfully",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/repositories/{repo_id}/merge": {
            "post": {
                "description": "Merges one branch into another branch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Git Operations"
                ],
                "summary": "Merge branches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Merge request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.MergeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branches merged successfully",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Merge conflict",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v2/repositories/{repo_id}/branches": {
            "get": {
                "description": "Retrieves all branches using V2 clean architecture with enhanced metadata",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "V2 Architecture"
                ],
                "summary": "List all branches (V2 Clean Architecture)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of branches",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "branches": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/api.BranchResponse"
                                    }
                                },
                                "count": {
                                    "type": "integer"
                                },
                                "current": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new branch using V2 clean architecture with enhanced branch management",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "V2 Architecture"
                ],
                "summary": "Create a new branch (V2 Clean Architecture)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Branch creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateBranchRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Branch created successfully",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.BranchResponse": {
            "type": "object",
            "properties": {
                "commit": {
                    "type": "string",
                    "example": "abc123def456"
                },
                "is_current": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "main"
                }
            }
        },
        "api.CheckoutRequest": {
            "type": "object",
            "required": [
                "branch"
            ],
            "properties": {
                "branch": {
                    "type": "string",
                    "example": "main"
                }
            }
        },
...
        "api.CreateBranchRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "from": {
                    "type": "string",
                    "example": "main"
                },
                "name": {
                    "type": "string",
                    "example": "feature/new-feature"
                }
            }
        },
...
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "REPO_NOT_FOUND"
                },
                "details": {},
                "error": {
                    "type": "string",
                    "example": "Repository not found"
                }
            }
        },
...
        "api.MergeRequest": {
            "type": "object",
            "required": [
                "from",
                "to"
            ],
            "properties": {
                "from": {
                    "type": "string",
                    "example": "feature/new-feature"
                },
                "to": {
                    "type": "string",
                    "example": "main"
                }
            }
        },
...
        "api.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Repository deleted successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter the token with the `Bearer: ` prefix, e.g. \"Bearer abcde12345\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ApiKeyQuery": {
            "description": "API key for accessing the API",
            "type": "apiKey",
            "name": "api_key",
            "in": "query"
        },
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}